requirements.txt
Flask
flask-cors
openai
google-genai
python-dotenv
requests